requirements.txt
streamlit
google-genai
pandas
requests
lxml
html5lib

